fsck.overlay: fix redirect feature when detect redirect dir xattr
The underlying layers may contian redirect xattr on some directories which
were renamed in redirect_dir mount options enabled overlayfs.

The old kernel do not set redirect feature on the upper layer when we mount
overlayfs with "redirect_dir=on", so this patch check and ask user to fix
this feature when we detect redirect xattr.

BTW, this patch also add redirect_dir feature to supported feature set, so
fsck will not fail if it detect redirect_dir feature in feature set xattr.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>